Pipeline to produce a forecast

Runs the pipeline with reasonable defaults and some hard-coded values to do the following. See the Examples. For now this function only works on Mondays!

Get data (national level from JHU by default)

Fit incident case and incident death models (ARIMA and lagged TSLM respectively)

Get future case data to create the incident death forecast

Create the incident death forecast based on this new data

Prepare submission format

Suggest a submission filename

Return all resulting objects to a list.

forecast_pipeline(
  method = "ts",
  source = "jhu",
  granularity = "national",
  horizon = 4,
  force = FALSE,
  ...
)

Arguments

method
Forecasting method to use; currently only 'ts' (time series) is supported

source
Data source to query; must be one of 'jhu' or 'nyt'; default is 'jhu'

granularity
Data aggregation level; must be one of 'national', 'state', or 'county'; if data source is 'nyt' then only 'national' can be used currently; default is 'national'

horizon
Horizon periods through which the forecasts should be generated; default is 4

force
Logical -- force the pipeline to run even if it isn't Monday? (default is FALSE; changing to TRUE may break something if today isn't Monday).

...
Arguments passed to other functions

Examples

if (FALSE) {
# Run all the steps to create models and forecasts
myforecast <- forecast_pipeline()
# Look at the submission and the suggested filename.
myforecast$submission
myforecast$submission_filename
# Write submission to file
readr::write_csv(myforecast$submission, file=myforecast$submission_filename)
# Validate submission
validate_forecast(myforecast$submission_filename)
}